feat: add public-types-lib separate npm package from domain-lib
