+++ packages/implementations-lib/package.json
@@ -31,18 +31,19 @@
     },
     "dependencies": {
         "@mojaloop/cert-management-bc-domain-lib": "*",
+        "@mojaloop/cert-management-bc-public-types-lib": "*",
         "@mojaloop/logging-bc-public-types-lib": "~0.5.4",
         "@mojaloop/account-lookup-bc-client-lib": "0.5.2",
         "@mojaloop/participant-bc-public-types-lib": "~0.5.4",
         "@mojaloop/participants-bc-client-lib": "~0.5.3",
         "@mojaloop/security-bc-client-lib": "~0.5.9",
         "@mojaloop/security-bc-public-types-lib": "~0.5.6",
-        "mongodb": "^6.4.0"
+        "mongodb": "^6.5.0"
     },
     "devDependencies": {
         "eslint": "^8.57.0",
         "jest": "^29.7.0",
-        "npm-check-updates": "^16.14.15",
+        "npm-check-updates": "^16.14.16",
         "ts-jest": "^29.1.2",
         "typescript": "^4.7.3"
     },
