+++ packages/mcm-internal-svc/package.json
@@ -39,6 +39,7 @@
     },
     "dependencies": {
         "@mojaloop/cert-management-bc-domain-lib": "*",
+        "@mojaloop/cert-management-bc-public-types-lib": "*",
         "@mojaloop/logging-bc-client-lib": "^0.5.6",
         "@mojaloop/platform-configuration-bc-client-lib": "^0.5.4",
         "@mojaloop/platform-configuration-bc-public-types-lib": "^0.5.2",
@@ -46,7 +47,7 @@
         "@mojaloop/security-bc-public-types-lib": "^0.5.6",
         "asn1js": "^3.0.5",
         "express": "^4.18.3",
-        "pkijs": "^3.0.15"
+        "pkijs": "^3.0.16"
     },
     "devDependencies": {
         "@mojaloop/logging-bc-public-types-lib": "^0.5.4",
