+++ packages/public-types-lib/.nvmrc
@@ -1,0 +1,1 @@
+20.10
