+++ packages/public-types-lib/LICENSE.md
@@ -1,0 +1,10 @@
+# LICENSE
+
+Copyright © 2017 Bill & Melinda Gates Foundation
+
+The Mojaloop files are made available by the Bill & Melinda Gates Foundation under the Apache License, Version 2.0
+(the "License") and you may not use these files except in compliance with the [License](http://www.apache.org/licenses/LICENSE-2.0). You may obtain a copy of the License at
+
+[http://www.apache.org/licenses/LICENSE-2.0](http://www.apache.org/licenses/LICENSE-2.0)
+
+Unless required by applicable law or agreed to in writing, the Mojaloop files are distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the [License](http://www.apache.org/licenses/LICENSE-2.0).
