+++ packages/public-types-lib/src/errors.ts
