+++ packages/public-types-lib/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+	"rootDir": "./src",
+	"outDir": "./dist"
+  },
+  "include": [
+	  "./src"
+  ]
+}
